Draw a five-pointed star
The file 'estrela-cinco-pontas.html' uses HTML5 and JavaScript to
draw a five-pointed star.

In order to draw such star, it is used two circumferences as
reference to draw each vertex of the star. Furthermore, it is used
polar coordinates to find out the distance from the center of the
star to each vertex. This approach enables to render the star in the
center of the canvas area. Besides, it is possible to draw different
stars just changing the radius of the circumferences.