Show the FPS in the screen
Show the Frames Per Second (FPS) in the game screen. It helps to
evaluate the performance of the game. The ideal is a FPS of 60. Lower then
that means the game is losing performance in that specific area, so you
should try to improve it in order to not have problems with collisions
detection, for instance.