Improve collision with the life drop
For some reason (probably Phaser bug), the callback function that handles
collisions with the life drop is called twice. To overcome this situation,
it is defined a flag in the life drop sprite, which is set to true when the
player collides with the life drop for the first time. This flag basically
disables the collision handler when it is set to true.