Fix collision with the hot sand
It is used the getTileWorldXY function to check whether the player
is over the hot sand (instead of trusting in Phaser to call the
group callback function to treat collision with the hot sand).

This approach also fixes the collision problem, which occurs when
the player is in "sun drop" state.