Fix camera movement when player is inside straw
Make the movement of the camera smoother when the player is inside the
straw. It uses the FPS to calculate how much the camera should move.
This approach works fine when the FPS is about 60. When the FPS is low,
the camera movement is not so smooth, but it is much better than the
previous solution.