Use hotkeys to make the drop jump
Using input.keyboard to jump makes the jump event happens several times.
This is not a problem for jumping now, but it will be when we change the
mechanic of the energy drop. Because of that, we replace the input.keyboard
by hotkeys. Using hotkeys, the action of jumping is detected only once,
that is what we want.