Enable the player to pause the game
Now the player is able to pause the game pressing Enter.

In order to implement this feature, we need to change how the timers of
the game are implemented. Instead of using the setTimeout function, we use
the Phaser.Timer class now. This change is needed because Phaser is not
able to pause and resume the timers when they are fired using setTimeout
function.