Bugfix: decrease two lives at once
Fix the bug of decreasing two lives at once when the player hits the hot
sand in normal state (drop is small).

Due to set the countCall variable to zero in the restartGame function, the
killDrop function is called twice. Hence, the life is decreased twice as
well. To overcome this problem, the countCall variable is now set to zero
in the create function.