Using the drop counter when creating character
Use the drop counter to decice the size of the drop. For example, if the
drop counter is equal to zero, then the Character class will create a
small drop. If the drop counter is bigger than zero, then the drop uses
the big size animation.

This change is needed to correctly update the drop frame when the player
pass from the first level to the second.