Bugfix: player hit hot sand and crab together
Attempt to fix a bug that happens when the player hit the crab, but the
timer of the hot sand is fired almost at the same time. This event causes
the player to lose two lives, instead of one.

This bug is very hard to reproduce, so it is not sure that this bug will
not happen anymore.