WIP: Add animation when the player loses
Play an animation and a sound when the player loses. The sound added is
temporary. The new sound will be added in other patch.

KNOWN ISSUES: it is not possible to pause and resume the song of defeat when
the player pauses the game. Phaser doesn't pause the song when the game is
paused. Besides, there is no event in Phaser that is fired when the sound
is completed, and this is needed to make this feature works.