Fix pause in the second level
This patch fixes a bug with the pause in the second level. The player
wasn't able to pause the game because of some conflits with the pause
implementation of the first level. In order to solve that, the pause
event of the first level is removed when the stage is clear.